// src/types/userType.tsx
export type UserType = {
  id: string;
  createdAt: string;
  nome_completo: string;
  tipo: string;
  telefone: string;
  email: string;
  senha: string;
  cpf: string;
  municipio: string;
  estado: string;
  bairro: string;
  endereco: string;
  complemento: string;
  avaliacoes: number;
  descricao: string;
  foto: string;
  profissao: string;
};